--- 4-2-timer/README
@@ -1,9 +1,0 @@
-Функция проверяет, имеет ли массив от 1 до 3 префиксов 'h', 'm', 's' (часы, минуты и секунды) и, если они есть, извлекает значения времени, связанные с этими префиксами. Затем она проверяет порядок префиксов, чтобы убедиться, что он корректен (например, 'hms' или 'hm', но не 'hsm'). Далее она проверяет значения времени, чтобы убедиться, что они находятся в допустимом диапазоне, а именно, что часы находятся в диапазоне от 0 до 23, а минуты и секунды находятся в диапазоне от 0 до 59. Если в процессе проверки возникнет ошибка, функция вернет false. Если все проверки успешны, функция вернет true и выведет сообщение "All validation is ok. Code - 201".
-
-Оценить уровень программиста по такому короткому фрагменту кода не всегда возможно, однако можно сделать некоторые предположения на основе качества кода.
-
-Код выглядит достаточно чистым и понятным, есть комментарии к коду, что является хорошим знаком. Функция имеет определенную задачу, и код в ней организован логично и структурированно.
-
-Однако, хороший код должен быть более универсальным и гибким, а здесь мы видим жестко заданные условия для проверки аргументов, и нет возможности добавлять новые условия или правила без изменения кода.
-
-Поэтому на основе только этого фрагмента кода сложно точно оценить уровень программиста, но, скорее всего, это может быть человек на уровне джуна или начинающего мидла.
